PhyloGenes functionality added
Using PhyloGenes means that given genomes, and a starting reference genome comprising of all the genes present in said reference genome, this tool can automatically generate a geneList.txt file, which is used to separate the tree by groupings. Includes a new parameter, "--phylogenes" to be used in conjuction with PhyloTree.sh if PhyloGenes.sh has been used.